+++ public/index.html
@@ -57,5 +57,6 @@
     <script src="https://unpkg.com/flowbite@1.4.7/dist/flowbite.js"></script>
 
     <script src="./TW-ELEMENTS-PATH/dist/js/index.min.js"></script>
+    <script type="text/javascript" src="https://cdn.omise.co/omise.js"></script>
   </body>
 </html>
