+++ src/components/pages/products/BuyingProduct.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function BuyingProduct() {
+  return <div>BuyingProduct</div>;
+}
+
+export default BuyingProduct;
