--- src/components/products/BuyingProduct.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function BuyingProduct() {
-  return <div>BuyingProduct</div>;
-}
-
-export default BuyingProduct;
